Refactor OptionSliderControl and use it in SimulationDialog
- Moved OptionSliderControl.tsx from blocks to parts to promote reuse
- Updated all import references to OptionSliderControl across the codebase
- Refactored SimulationDialog.tsx to use OptionSliderControl for Cycle and Player Num
- Ensured consistent UI and behavior for numerical inputs and sliders
- Verified that all lint checks and tests pass
