+++ src/components/parts/OptionSliderControl.tsx
