+++ src/feature/amongus/AuOptionControl.tsx
@@ -1,6 +1,6 @@
-import { OptionSliderControl } from "@/components/blocks/OptionSliderControl";
 import { OptionToggleControl } from "@/components/blocks/OptionToggleControl";
 import { OptionDropdownControl } from "@/components/parts/OptionDropdownControl";
+import { OptionSliderControl } from "@/components/parts/OptionSliderControl";
 import { translationMetaData } from "@/logics/api";
 import type { AuOptionMeta } from "@/type";
